Move app-only things from api to app, keep cmds decorators in single module
