+++ src/snowcli/cli/containers/compute_pool/commands.py
@@ -1,7 +1,6 @@
 import typer
-from snowcli.api.commands.decorators import global_options_with_connection
+from snowcli.api.commands.decorators import global_options_with_connection, with_output
 from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import CommandResult, SingleQueryResult
 from snowcli.cli.containers.compute_pool.manager import ComputePoolManager
