+++ src/snowcli/cli/containers/jobs/commands.py
@@ -2,9 +2,8 @@ import sys
 from pathlib import Path
 
 import typer
-from snowcli.api.commands.decorators import global_options_with_connection
+from snowcli.api.commands.decorators import global_options_with_connection, with_output
 from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import CommandResult, SingleQueryResult
 from snowcli.cli.containers.common import print_log_lines
 from snowcli.cli.containers.jobs.manager import JobManager
