+++ src/snowcli/cli/registry/commands.py
@@ -4,9 +4,8 @@ from typing import Optional
 import requests
 import typer
 from click import ClickException
-from snowcli.api.commands.decorators import global_options_with_connection
+from snowcli.api.commands.decorators import global_options_with_connection, with_output
 from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import CollectionResult, ObjectResult
 from snowcli.cli.registry.manager import RegistryManager
