+++ src/snowcli/cli/snowpark/procedure_coverage/commands.py
@@ -4,10 +4,10 @@ import typer
 from snowcli.api.cli_global_context import cli_context
 from snowcli.api.commands.decorators import (
     global_options_with_connection,
+    with_output,
     with_project_definition,
 )
 from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS, identifier_argument
-from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import CommandResult, MessageResult, SingleQueryResult
 from snowcli.cli.snowpark.commands import (
     get_app_stage_path,
