+++ src/snowcli/cli/streamlit/commands.py
@@ -9,10 +9,10 @@ from snowcli.api.cli_global_context import cli_context
 from snowcli.api.commands.decorators import (
     global_options_with_connection,
     with_experimental_behaviour,
+    with_output,
     with_project_definition,
 )
 from snowcli.api.commands.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import (
     CommandResult,
     MessageResult,
