+++ src/snowcli/cli/streamlit/streamlit_utils.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import List, Optional
 
 from jinja2 import Environment, FileSystemLoader
-from snowcli.api.utils.file_utils import templates_path
+from snowcli.cli.snowpark.package.utils import templates_path
 from snowcli.cli.snowpark.package_utils import parse_requirements
 
 log = logging.getLogger(__name__)
