+++ test_external_plugins/snowpark_hello_single_command/src/snowflakecli/test_plugins/snowpark_hello/commands.py
@@ -1,6 +1,5 @@
 import typer
-from snowcli.api.commands.decorators import global_options_with_connection
-from snowcli.api.output.decorators import with_output
+from snowcli.api.commands.decorators import global_options_with_connection, with_output
 from snowcli.api.output.types import CommandResult, SingleQueryResult
 from snowflakecli.test_plugins.snowpark_hello.manager import SnowparkHelloManager
