+++ tests/test_experimental_behaviour.py
@@ -1,9 +1,12 @@
 import json
 
 import typer
-from snowcli.api.commands.decorators import global_options, with_experimental_behaviour
+from snowcli.api.commands.decorators import (
+    global_options,
+    with_experimental_behaviour,
+    with_output,
+)
 from snowcli.api.commands.experimental_behaviour import experimental_behaviour_enabled
-from snowcli.api.output.decorators import with_output
 from snowcli.api.output.types import MessageResult
 from snowcli.api.plugin.command import (
     SNOWCLI_ROOT_COMMAND_PATH,
